+++ src/Prime/EventDispatcher/EventListener/NotificationLoggerListener.php
@@ -5,34 +5,69 @@ declare(strict_types=1);
 namespace Flasher\Prime\EventDispatcher\EventListener;
 
 use Flasher\Prime\EventDispatcher\Event\NotificationEvents;
+use Flasher\Prime\EventDispatcher\Event\PersistEvent;
 use Flasher\Prime\EventDispatcher\Event\PresentationEvent;
 
+/**
+ * @internal
+ */
 final class NotificationLoggerListener implements EventListenerInterface
 {
-    private NotificationEvents $events;
+    private NotificationEvents $dispatchedEnvelopes;
+
+    private NotificationEvents $displayedEnvelopes;
 
     public function __construct()
     {
-        $this->events = new NotificationEvents();
+        $this->dispatchedEnvelopes = new NotificationEvents();
+        $this->displayedEnvelopes = new NotificationEvents();
     }
 
     public function reset(): void
     {
-        $this->events = new NotificationEvents();
+        $this->dispatchedEnvelopes = new NotificationEvents();
+        $this->displayedEnvelopes = new NotificationEvents();
     }
 
-    public function __invoke(PresentationEvent $event): void
+    public function __invoke(object $event): void
     {
-        $this->events->add(...$event->getEnvelopes());
+        if ($event instanceof PersistEvent) {
+            $this->onPersist($event);
+        }
+
+        if ($event instanceof PresentationEvent) {
+            $this->onPresentation($event);
+        }
     }
 
-    public function getEvents(): NotificationEvents
+    public function onPersist(PersistEvent $event): void
     {
-        return $this->events;
+        $this->dispatchedEnvelopes->add(...$event->getEnvelopes());
     }
 
-    public function getSubscribedEvents(): string
+    public function onPresentation(PresentationEvent $event): void
     {
-        return PresentationEvent::class;
+        $this->displayedEnvelopes->add(...$event->getEnvelopes());
+    }
+
+    public function getDispatchedEnvelopes(): NotificationEvents
+    {
+        return $this->dispatchedEnvelopes;
+    }
+
+    public function getDisplayedEnvelopes(): NotificationEvents
+    {
+        return $this->displayedEnvelopes;
+    }
+
+    /**
+     * @return string[]
+     */
+    public function getSubscribedEvents(): array
+    {
+        return [
+            PersistEvent::class,
+            PresentationEvent::class,
+        ];
     }
 }