+++ src/Prime/Test/Constraint/NotificationOptions.php
@@ -30,7 +30,7 @@ final class NotificationOptions extends Constraint
             return false;
         }
 
-        foreach ($other->getNotifications() as $notification) {
+        foreach ($other->getEnvelopes() as $notification) {
             if (!array_diff_assoc($this->expectedOptions, $notification->getOptions())) {
                 return true;
             }
@@ -43,7 +43,7 @@ final class NotificationOptions extends Constraint
     {
         $actualOptions = [];
         if ($other instanceof NotificationEvents) {
-            foreach ($other->getNotifications() as $notification) {
+            foreach ($other->getEnvelopes() as $notification) {
                 $actualOptions[] = json_encode($notification->getOptions());
             }
         }