+++ src/Prime/Test/Constraint/NotificationType.php
@@ -40,7 +40,7 @@ final class NotificationType extends Constraint
             return false;
         }
 
-        foreach ($other->getNotifications() as $notification) {
+        foreach ($other->getEnvelopes() as $notification) {
             if ($notification->getType() === $this->expectedType) {
                 return true;
             }
@@ -60,7 +60,7 @@ final class NotificationType extends Constraint
     {
         $actualTypes = array_map(function (NotificationInterface $notification) {
             return $notification->getType();
-        }, $other->getNotifications());
+        }, $other->getEnvelopes());
 
         $uniqueTypes = array_unique($actualTypes);
         $typesList = implode(', ', $uniqueTypes);